Back-end/requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.6
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
whitenoise==4.1
xxx==0.0.4
djangorestframework=3.9.0
django-postgres-copy=2.3.7
